[FIX] l10n_be, l10n_it, l10n_se/17.4,18.0: set pos_receipt_label
When a client upgrades from v17.0 to v18.0, the pos_receipt_label field should
be populated automatically based on the company's localization. This update
addresses an issue where this field was not filled during the upgrade process.

The field was introduced in l10n_be and l10n_se in 17.4
and in l10n_it in 18.0

CoAs included in this upgrade are: be, be_asso, be_comp, it, se, se_K2, se_K3

related PRs:
https://github.com/odoo/odoo/pull/167012
https://github.com/odoo/odoo/pull/180458
https://github.com/odoo/enterprise/pull/74623

Task ID: 4274742

closes odoo/upgrade#6714

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>