[FIX] website_forum/18.0: force update of kanban arch
A new API for kanban templates has been introduced in 18.0 [1],
which deprecates the former API (the one defining a "kanban-box"
template). PR [2] adds a server-side warning when a legacy kanban
is met. In website_forum, there's a kanban view that is marked as
noupdate="1". This script forces the update of that view to the
new API.

This migration script is linked to [2] (community) and [3] (ent).

[1] https://github.com/odoo/odoo/pull/167751
[2] https://github.com/odoo/odoo/pull/184249
[3] https://github.com/odoo/enterprise/pull/72293

closes odoo/upgrade#6733

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>