[FIX] documents/saas~17.5: set the contact if we reset the owner
Purpose
=======
During the migration, the owner can be reset, if the user does not have
a write access on the documents
(because being owner gives all access on the record).

Because of that, the filter "My Documents" does not show most
of our files on next.odoo.com. To mitigate that, that filter now uses the
contact in addition to the owner, and during the migration, if we reset
the owner, and if the contact is False, we will set the contact to the
partner of the owner.

Task-4310779
See odoo/enterprise/pull/73357

closes odoo/upgrade#6730

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>