[IMP] documents/saas~17.5: hide user specific folders for internal users
Purpose
=======
When we create a document inside a folder,
it inherits by default the access of the folder.

We want the newly created documents to not be accessible by everyone,
when they are created in an old user specific folder, readable by internal
users, and so we decided to restrict the access of those folders
(the logic of the files inside of it stay the same).

Task-4307017

closes odoo/upgrade#6710

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>