[IMP] l10n_no/saas~18.1: add standard tax code
For correct reporting of the SAFT report in Norway, a standard code for
taxes is used. Until now, it was only found in the description or in the
tax name (in earlier version). For more robustness, the code is now
added as a field in the Norwegian localisation.

Linked Community PR: https://github.com/odoo/odoo/pull/186745

task-3975794

closes odoo/upgrade#6742

Related: odoo/odoo#186745
Related: odoo/enterprise#73336
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>