[FIX] documents(_spreadsheet)/saas~17.5: keep access_internal if we can
Purpose
=======
During the migration, a folder can't be set as access_internal=edit,
to not give a write access on it to all users.

```
access_documents_folder_base_group_user,...,base.group_user,1,0,0,0
```

But, if that folder is pinned, normal users won't be able to write in it
(eg rename, etc), and so we decided to keep access_internal=edit for the
user_specific folders (eg the "Spreadsheet" folders), so user can still
upload in it after the migration.

Task-4328340

closes odoo/upgrade#6751

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>