[FIX] account/saas~13.2: correctly update sequence_number column
Since odoo/upgrade#6621 it's possible that the sequence_number
column is NULL for some rows in the account_move table. That's coming
from the double '{' in the second parameter of the "default match"
query. The call to query.replace will turn these '{{' into '{{{{'
before executing the query, leading to NULL values for the sequence_number
of these rows.

This leads to a performance bottleneck when posting invoices
in a database that got upgraded from 13.0 up to 16.0, 17.0 and 18.0.
Because of the empty sequence_numbers, the computation of an account.move
name will retrieves the first move with an empty sequence and use that
as the last_sequence to increment.

Example, if for a given journal_id, moves named STC/2024/0001
up to STC/2024/23000 have empty sequence_numbers, the next sequence
computation query will fetch STC/2024/0001 as the last sequence number.
Then it will increment it, giving STC/2024/0002, before trying to insert
the result in the database. As STC/2024/0002 already exists, this will
raise a UniqueViolation error and Odoo will retry with STC/2024/0003,
failing every time until STC/2024/23001. This can produce a huge number
of queries, leading to slowdowns when posting invoice.

After this fix, no sequence_number should be NULL when upgrading from 13.0.
Also, the behavior after this fix is consistent with the behavior of the
migration script before odoo/upgrade#6621

opw-4328377

closes odoo/upgrade#6767

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>