[IMP] base/saas~18.1: check indexes and WHERE bool
Warn about indexes that may use a boolean field in their
condition because the generated queries changed for boolean fields.
The ORM will translate queries on booleans as `field IS [NOT] TRUE`
instead of comparisons with a boolean value or null checks.
We want to identify indexes that may not be useful anymore:
these contain a WHERE clause with a comparison with a boolean value
or "active" field while ignoring conditions that use the new syntax.

Remove indexes that will be recreated by the ORM with a different name.

task-4067946
odoo/odoo#187538

closes odoo/upgrade#6776

Related: odoo/enterprise#73981
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>