[FIX] mrp/saas~17.2: support old manufacture rules behavior
Follow-up on 61c216b.

The `manufacture` rules case was overlooked, and potentially breaks
custom routes silently after upgrading. This applies the same fix as the
previous commit, but for `manufacture` rules as well.

The reason is the same: We want to be able to keep the old behavior as
well, being able to set a destination on a rule and have it supersede
the one defined in the linked picking type.

opw-4300074
See odoo/odoo#187898

closes odoo/upgrade#6787

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>