[IMP] account_reports/saas~17.3: don't delete mat tax report
In PR odoo/enterprise#62890 the external identifier
account_reports.mail_activity_type_tax_report_to_be_sent deleted while
it is an important feature.

PR odoo/enterprise#73615 reintroduced it.

As the record is the same, we just don't delete it anymore.

Note:
For the sack of testing or in the hypothetical case someone would
upgrade to saas~17.3, we would delete the record as it doesn't exist in
the dead saas~17.3 version.

Task-4292304

closes odoo/upgrade#6788

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>