[FIX] documents/saas~17.5: Create tags and server actions in post script
The server actions reference records with `forcecreate=0`, so to avoid issues
while updating the module, they should be marked as noupdate.

They can then be created in a post script directly from the data file.

This also avoids an issue with the current script, that creates the records
through a query, but it violates the `NOT NULL` constraint on `facet_id`.

tbg-1641

```
psycopg2.errors.NotNullViolation: null value in column "facet_id" of relation "documents_tag" violates not-null constraint
```

Part-of: odoo/upgrade#6803
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>