[FIX] base/0.0.0: do not call _check_xml in `write`
In the linked PR we now call `_check_xml` in `write` when activating
views. Since the views fixer extensively use activation and
de-activation of views we need to defuse the check on `write`. All views
are fully checked in this script therefore there is no need for any
extra check in `write`.

See odoo/odoo#187607

closes odoo/upgrade#6809

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>