[IMP] project/saas~18.1: replace `display_in_project` field with filter
If subtasks (tasks with a `parent_id`) had the same project as their
parent, we used an boolean icon that, if (un)checked, made the task
(in)visible in the project.
We removed that field from the view, and made all subtask that have the
same project as their parent invisible. Instead, we created a
"pseudo-filter" that displays them if needed.
So, there's no need for the field `show_display_in_project` anymore,
because we'll never display it.

This commit also alters the `display_in_project` field for the subtask having
the same project than the parent task because since the changes made by
the related community PR, it is no longer possible for the user to edit
the `display_in_project` field. Before the changes made, the user was
able to display or not the subtask with same project than the parent
task but now, the `display_in_project` field will only be a technical
field.

And so, now, those subtasks will no longer be visible in the main tasks
view of the project. The user will have to select the new filter `Show
sub-tasks` to see show them.

Community PR: odoo/odoo#185912
Enterprise PR: odoo/enterprise#73124

task-4295713

closes odoo/upgrade#6811

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>