[FIX] documents/saas~17.5: fix queries
When using format_query for identifiers like tables
and columns and fkeys the double quotes surrounding
the names are added automatically, so the input
variables should not be already quoted otherwise
we get wrong names in the query.
In this part of the script the custom fkeys may
be quoted in the result of util.get_fk.

closes odoo/upgrade#6839

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>