[FIX] account/saas~17.3: invoice_currency_rate update for cz/hu
In 17.3 the currency rate was stored and displayed on the
'account.move' level.
See commit df264fd8ef68120d1207ad67e13fd505289f0cae.

For modules `l10n_cz` and `l10n_hu_edi` the date used for the
currency conversion is currently different than "standard".
  * For `l10n_cz` we use the Taxable Supply Date (if it is set)
  * For `l10n_hu_edi` we use the Delivery Date (if it is set)

The forward-ports introducing the changes for the localizations
were not adapted for the change in 17.3 mentioned above.
This was recently corrected for community in odoo/odoo#189023.

This commit fixes the upgrade script to use the correct dates.

See odoo/odoo#189023

closes odoo/upgrade#6853

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>