[FIX] hr_expense_extract/saas~17.2: remove mixin
In module 'hr_expense_extract' the account.move model was
overwriting the '_needs_auto_extract' method of the extract
mixin, assuming that it was already inherited by the model.
The inheritance is available not by direct dependency but
by the 'account_invoice_extract' module being autoinstalled.
After a refactor, the overwrite is no longer needed in
hr_expense_extract, which means that the account.move model
does not inherit the mixin anymore unless the other module
is installed.
The change in enterprise affected versions from 17->saas~17.2
but since the mixin started getting inherited only in 17 then
the change will only be visible when passing by saas~17.2

see https://github.com/odoo/enterprise/pull/49816

closes odoo/upgrade#6855

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>