[IMP] stock/tests: skip some products' qty check
There have been changes in the way quantity on hand is computed in MRP.
The changes impact how the values are rounded in intermediate
computations that later affect the final value. There is no way we can
make versions coincide because fundamentally the rounding would always
have an edge case.

Thus, in this patch we propose to allow the skip of some product checks
but only if the difference is small enough.

closes odoo/upgrade#6856

See: odoo/odoo#189605
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>
Co-authored-by: sagu-odoo <sagu@odoo.com>