[IMP] spreadsheet_edition/saas~18.1: improve pivot sorting
With the o-spreadsheet commit, we allow to sort any pivot on a column
(spreadsheet pivots in addition of odoo pivots). We thus changed the
`sortedColumn` of the pivot definition from web by our own
`sortedColumn`.

This commit migrate the `sortedColumn` of the pivot definition. Only
the pivot sorted on the total column are migrated, as we're missing
some information to migrate the others. We remove the `sortedColumn`
from the pivots that are not sorted on the total column.

odoo/o-spreadsheet#4998
odoo/odoo#191109

closes odoo/upgrade#6873

Task: 3989395
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>