[FIX] account/saas~16.2: fix tax group duplication
In the case where we have multiple modules with
different tax groups for different countries, we
want to recreate the tax groups with a company_id
only for the companies that can use them.
The way it was happening here is by copying the
groups for any company that is associated with
any tax group.
A bigger problem as a consequence of having all
those copies with all companies, is that when it
is time to create xmlids, we can end up with a
wrong mapping, because all xmlids have module
'account' and the name may be the same across
different localizations, so the only difference
should be the company_id prefix. In the current
setup we choose the first one and skip the
possibly correct tax group for the company due to
the 'on conflict do nothing'.

closes odoo/upgrade#6877

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>