[IMP] sale_subscription/saas~18.1: set effective date on log
Before this commit, all logs were used to compute MRR. As a result it
was not possible to filter out logs not invoiced yet.

During upgrade, we decide to set the effective date equal to the event
date to ease the script. As a result, the effective dates are not
correct for not yet invoiced subscription but it's not a real issue.

closes odoo/upgrade#6893

Taskid: 4391586
Enterprise: https://github.com/odoo/enterprise/pull/75376
Related: odoo/enterprise#75376
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>