[FIX] base/saas~17.5 : Update Modified Language
Issue:
------
Invalid language code 'sr_RS' while creating an invoice.
Cause:The language code of one of the serbian language 'sr_RS' has been modified
to 'sr@Cyrl' and is not updated for the partner's lang field which is character.
So, after upgrading to 18.0 when trying to create an invoice using the partner
having serbian language with code 'sr_RS' will raise an error.

Solution:
-----------
Updated the lang field value to the new code which is 'sr@Cyrl'.

Steps to reproduce:
------------------------
1. Create a database in lower version than 18.0 and install serbia lang 'sr_RS'.
2. Create a new partner/ a user and set Serbian language('sr_RS') in the
   lang field.
3. Upgrade to 18.0 and try creating a record using that partner.
   For ex: invoice
4. Will raise invalid language code 'sr_RS' error.

After the PR:
Community PR:
1.
https://github.com/odoo/odoo/commit/a05a8ee736a5e97d10138096ec1bd73419591901#diff-bf3cff12b0826b1864f337382ed26f80010674195d32a8cb8858a50ee20a5e2dR65

closes odoo/upgrade#6912

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>