[FIX] base/saas~17.5 : Update url-code value
Issue:
------
Invalid language code 'sr_RS' in the url after migrating the database to 18.0
from lower versions.
Cause: After modifying the language code to 'sr@Cyrl' from 'sr_RS', the url_code
should also be updated as the lang code or its iso_code. But the deprecated one
still remains in the url as the url_code when switching to the serbian language.

Solution:
-----------
Updated the url_code value to the new one which is 'sr@Cyrl'.

Steps to reproduce:
------------------------
1. Create a database in lower version than 18.0 and install serbia lang 'sr_RS'.
2. Upgrade to 18.0 and switch to the other language i.e; serbia
3. Look into the url, the old url_code 'sr_RS' remains.
4. Or check languages menu, there the url_code will not be updated.

REF PRS:
Community PR:
1. https://github.com/odoo/odoo/commit/a05a8ee736a5e97d10138096ec1bd73419591901#diff-bf3cff12b0826b1864f337382ed26f80010674195d32a8cb8858a50ee20a5e2dR65
2. https://github.com/odoo/odoo/blob/35828b161b5c643c4a9554d3c6a28e5dcf614e00/odoo/addons/base/models/res_lang.py#L288

closes odoo/upgrade#6919

Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>