[FIX] project/tests: ensure personal stages exist
The test case TestMigrateNoteInTodo in project requires
the env user to have at least 1 personal stage before
the execution of the test_prepare. This condition is not
always met based on the order of running of other test
cases. The test_check also ensures that the number of
personal stages is the same between multiple users,
so here we propose creating a dummy task to trigger
the creation of any missing standard personal stages

closes odoo/upgrade#6922

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>