[ADD] loyalty/saas~17.5.1.0: Set `loyalty.card` as active
The field `active` on the model `loyalty.card` was added in odoo/odoo#171706
without any proper script to fill a value for it.

Databases which upgrade from a version lower than 18.0 don't have any
value set for this field which leads to some confusion for customers
who think there is a loss of the records that were present before.

closes odoo/upgrade#6894

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>