[ADD] l10n_mx_edi{,sale}/saas~18.1: Compute payment policy
In the related PR, we changed the payment policy field to a stored
computed field.
We manually compute `l10n_mx_edi_payment_policy` on account.move so that
it can be done efficiently (exploded).

Note that the `partner_id.l10n_mx_edi_payment_policy` field (used in the
compute method in the related PR) defaults to False, so we don't need to
reflect that condition in the query.

We also introduced a new stored computed field for the payment policy on
sale.order, which defaults to False, so we create the column in DB with
NULL values.

closes odoo/upgrade#6938

See: https://github.com/odoo/enterprise/pull/75709
Related: odoo/enterprise#75709
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>