[IMP] website_sale/saas~18.3: remove `cart_item_line_heading`
See odoo/odoo#190720, the `website_sale.cart_item_line_heading` template
is removed because we handle the link condition on the `<a>` to apply a
list-group styling encapsulating the title and description of the combo
line.

These changes were made to simplify the cart_lines template and ensure
responsiveness and readability of the products on the checkout cart.

Related: odoo/odoo#190720
Related: odoo/enterprise#75679
task-4387239

closes odoo/upgrade#7446

Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>