[FIX] l10n_es/saas~16.5: fix constraint error of deprecated account tag
Some account tags are removed from version saas~16.5, but if a customer has
used that tag in the tax that is used in account_move_line then a record will be
stored in 'account_account_tag_account_move_line_rel' relation table. While
upgrading the database to version 17, the database was blocked with the error of
ForeignKeyViolation.

To resolve this issue, I added new tags that were recently added to the ES
localization to make a smooth upgrade.

Related PR: odoo/odoo@d821db7
upg-2402334
opw-4429878

closes odoo/upgrade#6985

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>