[FIX] web_studio/18.0: Add currency_id element
When upgrading to Odoo 18, if the customer modified with studio a report
that calls the report account.document_tax_totals_template,  or calls a
report inheriting from it, like: account.document_tax_totals,
sale.document_tax_totals, purchase.document_tax_totals),
the report fails with a traceback after the upgrade.

This is because of the way Studio deals with reports since 16.4
(copying the entire arch of the report into a new view).

On Odoo 18, the report account.document_tax_totals_template receives a
new parameter currency, which was not there on previous versions.
Raising the exception when not there

[here](https://github.com/odoo/odoo/blob/2fe55ae592e5812c5ee5e39ed1e8332bde608cb2/addons/account/views/report_invoice.xml#L189-L192)
opw-4418042
upg-2404122
unicorn-4417707

closes odoo/upgrade#6994

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>