[FIX] hr_expense/16.0 : Create attachment records for hr expense sheet
Upgrade the database from 15.0 (or older version) to 17.0,
In v17.0, When attempting to click on an expense sheet line (with an attachment)
for an existing record (created in version 15.0 or older), error occurs.
The error arises because the system is unable to retrieve the 'checksum' value
due to the absence of an attachment in the hr.expense.sheet model.

To resolve this issue, I created a copy of the hr.expense attachment,
updated the res_id and res_model fields accordingly,
associated it with the hr.expense.sheet model.

upg-2394352
opw- 4329027

closes odoo/upgrade#7004

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>