[FIX] base/0.0.0: fixed inconsistency for onhand qty
After the commit f0d1b4b5d0c6a13137349f0bcd27bba45f56ac46,
An inconsistency in the product on-hand quantity has been observed
Due to the change in how companies and warehouses are handled
in the upgrade from Odoo 17.0 to 18.0.

In version 17.0, the customer database only had one allocated company,
which limited the system to fetching the warehouse and related
internal and virtual locations of that specific company
when calculating product quantities. However, after the upgrade,
we are adding all companies to the root user, which causes the system
to fetch warehouses and locations for all companies,
leading to discrepancies in the on-hand quantities.

To resolve this issue, I moved the relevant logic into the init- file,
ensuring that the system will fetch product quantities in a consistent manner,
avoiding any conflicts caused by the expanded company data post-upgrade.

see:
https://github.com/odoo/odoo/blob/470a14092d2f8a0a1197f2f7665fccc6cd80cc46/addons/stock/models/product.py#L303

upg-2408289
upg-2408288
Upg-2408207

TGB:
https://upgrade.odoo.com/web#id=2&cids=1&menu_id=107&action=178&model=upgrade.request.traceback.group&view_type=form

closes odoo/upgrade#7012

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>