[FIX] industry_fsm_report/0.0.0: update depricated bootstrap
we upgraded from Bootstrap v4 to v5. As a result,
some classes have been deprecated and no longer work as expected.

https://github.com/odoo/odoo/pull/95450

For example, in Bootstrap v5, the CSS class font-weight-bold was
replaced with fw-bold. Therefore, font-weight-bold no longer
functions correctly.

NOTE: While we have handled the conversion of Bootstrap classes,
but condition is not applicable for the worksheet report.

https://github.com/odoo/upgrade/commit/cfd58d145b8bbb99b0ec1dcc70bcfbfa3f9b7e20
https://github.com/odoo/upgrade/blob/7b06e5b4f5f0be31f8d010b3f12f511e1584e59b/migrations/base/16.0.1.3/pre-90-convert-bootstrap5.py#L81-L103

Steps to Reproduce:
1. Create a demo database in version 15.0.
2. Install the industry_fsm_report module.
3. Navigate to Field Service > Configuration > Worksheet Template.
4. Create a new template (a new custom report
   will be generated for this template).
5. Assign the newly created template to a task.
6. Migrate the demo database to version 17.0.

In the upgraded database, the field label does not render in bold
for worksheet report.

Enterprise PR:
https://github.com/odoo/enterprise/pull/76509

closes odoo/upgrade#7020

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>