[FIX] documents/saas~17.5: Consider translated name for document folder
During the document refactor, we created the documents_document
record to use it as a folder.
https://github.com/odoo/upgrade/blob/b21e18a88827588e704fc222959eebcc69093ca1/migrations/documents/saas~17.5.1.4/pre-migrate.py#L80-L121
However, we only considered the name->>'en_US' for the name
value of new records.This results in incorrect folder names
after an upgrade if customers are using languages other
than English (en_US).
https://github.com/odoo/upgrade/blob/b21e18a88827588e704fc222959eebcc69093ca1/migrations/documents/saas~17.5.1.4/pre-migrate.py#L109

Steps to Reproduce:
- Create a demo database in version 17.4.
- Select a different language for the company (other than en_US).
- Navigate to Documents and create a new folder.
- Edit the folder and assign a name.
- Migrate to version 18.

Issue:
After the migration, the folder names differ from the ones
originally assigned. Instead, only the en_US name is used
for new records because it is hardcoded in query.

closes odoo/upgrade#7044

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>