[FIX] l10n_sa_edi/saas~17.5: encode private key
We need to ensure that the private key stored in
"res_company.l10n_sa_private_key" is encoded in base64
prior to being moved from "res.company" to "certificate.key".

When a private key for l10n_sa_edi is generated in Odoo 17,
it is not encoded in base64:
https://github.com/odoo/odoo/blob/20cd56a3442e5bf7d579b0be6f5c14d2c92b05a5/addons/l10n_sa_edi/models/res_company.py#L12-L27

This is contrary to other l10n modules, for example:
https://github.com/odoo/enterprise/blob/b026f6c76385caac70952c7e31fdf662b0aee1f2/l10n_ar_edi/models/res_company.py#L175

It is also contrary to design principles of Odoo:
https://github.com/odoo/odoo/blob/3f2f807a40abcabd9d8dce50bd4e034aaeb1fab3/odoo/orm/fields_binary.py#L56-L61

We ensure that it is encoded in base64 also because "certificate.key"
expects all keys stored in the "content" field to be encoded.
Otherwise when computing the "pem_key" field, the key is not
going to be retrieved properly from "content":
https://github.com/odoo/odoo/blob/3f2f807a40abcabd9d8dce50bd4e034aaeb1fab3/addons/certificate/models/key.py#L74

closes odoo/upgrade#7050

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>