[FIX] account_online_synchronization/0.0.0: Cron on demand
This commit ensures that the cron which fetches transactions asynchronously
is correctly deactivated like the xml data by updating the record from the xml.
This cron is activated when needed and only when it's needed.

We're forcing this deactivation because we don't want to run the cron every
5 minutes if no account online account is set on the database.

See https://github.com/odoo/enterprise/pull/77013

closes odoo/upgrade#7058

Related: odoo/enterprise#77013
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>