[FIX] account/saas~17.3: fix invoice_currency_rate update
Currently an error may happen in a multi-currency setup when computing
the invoice currency rate in case the total amount of the invoice is 0.

The issue is during the building of the temporary table
`_upg_currency_rates`. For each needed (company, currency) pair it
contains rates (from table 'res_currency_rate') for each day in a date range.
The date range starts with the first day on which we need a rate for
the (company, currency) pair and ends (inclusively) with the last day
on which we need such a rate.
To determine the rate on the first day we may have to look for the
last rate for the (company, currency) pair before the start of the
date range.
Currently it only filters by company though; so we may fetch a rate
from a different currency.

This commit fixes this problem.

Part-of: odoo/upgrade#6853
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>