[IMP] documents/saas~18.2: rename internal as administrative folder
We rename the internal folder as administrative folder as internal doesn't mean
much. We have decided to not rename the xmlid as it is used in the code of the
server action which is complex to update.

See odoo/enterprise#76080

Task-4294237

Part-of: odoo/upgrade#6820
Related: odoo/odoo#191459
Related: odoo/enterprise#74571
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>