[FIX] delivery/saas~16.1: correct shipping_weight compute query
Currently during the upgrade when the field shipping_weight is added to
the sale order table we are computing the values for the field in the
script upgrade/migrations/delivery/saas~16.1.1.0.

But in the query we are joining sale_order_line and sale_order table,
because of that in the table that is created "so", we have records for
each sale order line and when we try to update the sale order table
field shipping weight it just selects the first record from the
"so" table.

In order to fix the issue with the computation, I added a group by to be
on the sale order id and then used an aggreate function to calculate the
sum based on all the sale order lines.

closes odoo/upgrade#7086

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>