[FIX] sale/tests: fix order product_id
The saas~16.2 test for 'amount to invoice' creates a
product with explicitly no taxes to compare the
amount to invoice to price - downpayment with literal
numbers. The product_id used on the sale order line is
actually the product template id, so it is matching
another product with non zero taxes, leading to a check
failure '505.0 != 400.0'.

closes odoo/upgrade#7104

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>