[IMP] documents/saas~18.2: add sequence and make name translatable
The associated enterprise PR reintroduces drag-and-drop sorting of
folders in the search panel. To do this, we add a sequence field to
'documents.document'.

Before Documents' makeover (18.0), it was possible to translate
folder names.

To reintroduce that feature, and as now folder and documents share the
same model, we make 'documents.document' name field translatable.

see https://github.com/odoo/enterprise/pull/77890
task-4512349

closes odoo/upgrade#7133

Related: odoo/enterprise#77890
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>