[FIX] {base,product}/saas~14.5: ensure groups
If there are multiple currencies active we must set the Multi-Currency
group. If this group is set then we must also set the Pricelist group.

All this logic is encoded in `_toggle_group_multi_currency` and the
overrides of `_activate_group_multi_currency` (in `product`,original in
`base`). When loading `base/data/res_currency_data.xml` the toggle is
triggered due to new currencies like CNH, but the override in product is
not. We thus ensure here that the toggle is called once all modules are
loaded. This ensures that both the multi-currency and pricelist groups
are properly applied.

Steps to reproduce the issue:
1. In 14 install sale_management
2. In settings toggle multi-currency (note that this also activates
   pricelists)
3. Check that EUR and USD are both active (Invoicing > Settings >
   Currencies)
4. Deactivate multi-currency, and pricelists (note that both USD and EUR
   are kept active).
5. Upgrade to 15.

In the upgraded DB we get that multi-currency group is set, but not the
pricelist group. Whether if we would activate any other currency
directly in 15 the pricelist group would then be set.

See odoo/odoo#74396

closes odoo/upgrade#7128

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>