[FIX] sale_project_forecast/saas~16.3: deactivate plan services
In version 16, it was possible to have
the "Plan services" option activated for
products with invoice policy based on delivered qty (manual).

In version 17, as per sandard checkbox should be hidden
and also deactivated when upgrading to 17.0

here issue is - checkbox for `Plan Services` is hidden but
not deactivated, it's still active and will check if we
enable it by studio.
so need to correct it by making false value when
service policy is `Based on Milestones`
and `Based on Deliveryd Quantity (Manual)`

service policy is non store field so we need to compute
the field `plan services`
see:

https://github.com/odoo/enterprise/blob/saas-16.3/sale_project_forecast/models/product_template.py#L13

upg-1610272

unicorn - 4505581

closes odoo/upgrade#7131

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>