[FIX] hr_payroll_holidays/saas~18.2: prevent payslip stage reset
In the tree view of the payslip model, we can do mass edit to create a journal
entry. Before, we were able to perform this action on a paid payslip which does
not make sense. Now we have an error message if we try to do it.

Task: 3874129

See odoo/enterprise#77677

closes odoo/upgrade#7167

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>