[FIX] stock/saas~18.1: skip hour/day uom constraint
In saas~18.1 the work time uoms are forced to make
hours the reference unit and days relative to it,
in contrast to the previous versions where the
reference was the days.
If we have stock module installed this write may
fail because of a restriction on changing uom
factors for open stock moves.
Since the intention is to swap reference and not
actually change factors, we can skip the validation
if the write is coming from the uom end script

closes odoo/upgrade#7179

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>