[IMP] account_online_payment/saas~18.2.1.0: Adapt computed field name
The aim of this commit is changing the computed field
`account_online_linked` to `account_online_link_payments_enabled`.
We are doing this change because the previous behavior wasn't sufficient
for what we want to do with this field.
Now, this field is checking if we have payment activated on the
bank sync connection or not. To have a field name meaningful and
crystal clear, we decide to rename the field as well.

task-4532157

See also https://github.com/odoo/enterprise/pull/78740
See also https://github.com/odoo/odoofin/pull/388

closes odoo/upgrade#7194

Related: odoo/enterprise#78740
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>