[IMP] sms/saas~18.5: uniformize blacklist option name
We uniformize the name of the option that activate the use of the exclusion
list when sending sms or emails. Concretely, we rename the sms composer field
"mass_use_blacklist" to "use_exclusion_list" to match the name used in the mail
composer for the same feature.

See odoo/odoo#187403
See odoo/enterprise#93060

Task-3575361

Part-of: odoo/upgrade#7208
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>