[IMP] purchase/saas~18.2: b2b refresh UI
purchase_requisiont adds a useless button in tree view, we removed it.
We also remove the bill control settings since the behavior is now
on deliverd for all goods and ordered on services

Remove the mail confirmation and decline. It will be replaced by a
simpler acknowledgment flow. We only update the untouched template.
It won't be an issue if people still having the button in their template
since the route still exist and is backward compatible

closes odoo/upgrade#7213

Task: 4321341
Community: odoo/odoo#197211
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>