[FIX] l10n_in/saas~17.1: remove forcing the audit trail
The field "check_account_audit_trail" in res.company model was forced to true
if the company fiscal country was India. While it's not the case if a new
Indian company was created in a V17.1+ DB. So it's not right to force this
feature for the customers that upgraded their DB if they don't want it on.

closes odoo/upgrade#7216

Signed-off-by: William André (wan) <wan@odoo.com>