[FIX] website/0.0.0: remove website_id reference for internal user
when a user becomes an internal user and they have a website id
the  website id was kept which restricted
the user to sign in with one website. When an internal user has a
website id, they can't sign in with another website. This restriction
only makes sense for portal users and it confused odoo customers (we had
support tickets about that) so we decided to prevent having a website
id for internal users.

so to make upgrade successfull remove website reference from internal
user

we found issue on saas-17.4 to 18.0 too so better
move on 0.0.0 script

upg-2437868
upg-2559396

TGB:
https://upgrade.odoo.com/web#id=1453&cids=1&menu_id=107&action=178&model=upgrade.request.traceback.group&view_type=form

closes odoo/upgrade#7234

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>