[FIX] hr/saas~16.4: Avoid creating duplicate contacts
Here is what we want to achieve:
In version saas~16.4, it selects the records of hr_employee and search for
partner with the same email set in work_email of it. If the partner is searched
the it will set it to work_contact_id and if not then create a  new partner and
then set it to work_contact_id

Here are our cases:

Case 1: Employee active & Partner active
It will search and get the partner in search and set it to work_contact_id.

Case 2: Employee active & Partner inactive
It will not search as the partner is inactive hence create a new partner. In our
fix 'with_context(active_test=active, default_active=active)' the active will be
true and hence it will search for active partner and as the partner is inactive,
it will create a new partner which is active and link that new partner to the
work_contact_id.

Case 3: Employee inactive & Partner active
It will search and as the partner is active, it will be searched and hence set
to the work_contact_id.

Case 4: Employee inactive & Partner inactive
Partner will not get searched and hence will try to create a new partner. But
for resolution added 'with_context(active_test=active)' which will search for
its inactive record too and then, it will be set on the work_contact_id.

Case 5: Employee active & Partner deleted
Normally it will create a new partner which is active. And with fix too as I
have added `with_context(default_active=active)` then active is true in this
case so an active partner will be created and link that new partner to the
work_contact_id.

Case 6: Employee inactive & Partner deleted
Normally it will create a new partner which is active. And with fix too as I
have added `with_context(default_active=active)` then active is false and hence
an inactive partner will be created and link that new partner to work_contact_id

closes odoo/upgrade#7248

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>